Build: Export bookmarklets in the Netscape format
Instead of manually adding bookmarklets one by one, we can export a
'bookmarks.html' file in the standard Netscape bookmarks format. Modern
browsers typically use SQLite or JSON to store bookmarks internally, but
the HTML format is the only supported *interchange* format.

The HTML-based format goes back to the days of Netscape! It's kind of a
cursed format, but it's not that complicated, so we only need ~90 LOC to
support it (minus proper escaping which I'll add next).