Google: Manually specify the translation language
Letting Google guess the language to translate to is convenient, but
it introduces an additional history entry as Google Translate updates
the current URL when it determine's the user's preferred language. The
consequence is that the browser's back button has no apparent effect
unless it is pressed twice in quick succession.