REPL: Display an error message if `inspect` throws
We wrap the `inspect` call in its own `try...catch` statement in case it
throws an exception, e.g. by attempting to recursively inspect a cyclic
data structure. This is better than quitting without any explanation.

If displaying that error causes *another* exception to be thrown, the
original `catch` block executes as if `eval` threw an exception.